# server/docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.Facility:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  domain.FacilityCreatePayload:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  domain.FacilityListRes:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Facility'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/domain.Meta'
      status:
        type: integer
    type: object
  domain.FacilitySuccessRes:
    properties:
      data:
        $ref: '#/definitions/domain.Facility'
      message:
        type: string
      status:
        type: integer
    type: object
  domain.FacilityUpdatePayload:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
...
  domain.Meta:
    properties:
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  domain.SuccessRes:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A comprehensive API for school management system.
  license:
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: School Management System API
  version: "1.0"
paths:
...
  /facility/{facility_id}:
    delete:
      consumes:
      - application/json
      description: Delete facility
      parameters:
      - description: facility_id
        in: path
        name: facility_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: facility
          schema:
            $ref: '#/definitions/domain.SuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Delete facility
      tags:
      - Facility
    get:
      consumes:
      - application/json
      description: Get facility by id
      parameters:
      - description: facility_id
        in: path
        name: facility_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: facility
          schema:
            $ref: '#/definitions/domain.FacilitySuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Get facility by id
      tags:
      - Facility
  /facility/create:
    post:
      consumes:
      - application/json
      description: Create facility
      parameters:
      - description: facility
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.FacilityCreatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created new facility
          schema:
            $ref: '#/definitions/domain.FacilitySuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Create facility
      tags:
      - Facility
  /facility/list:
    get:
      consumes:
      - application/json
      description: Get facility list
      produces:
      - application/json
      responses:
        "201":
          description: facility
          schema:
            $ref: '#/definitions/domain.FacilityListRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Get facility list
      tags:
      - Facility
  /facility/update:
    put:
      consumes:
      - application/json
      description: Update facility
      parameters:
      - description: facility
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.FacilityUpdatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Updated facility
          schema:
            $ref: '#/definitions/domain.FacilitySuccessRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - JWT: []
      summary: Update facility
      tags:
      - Facility
...
schemes:
- http
- https
securityDefinitions:
  JWT:
    description: JWT security accessToken. Please add it in the format "Bearer {AccessToken}"
      to authorize your requests.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"