# server/domain/facility.go
package domain

type Facility struct {
	Base
	Name string `json:"name" binding:"required"`
}

type FacilityCreatePayload = Facility
type FacilityUpdatePayload = Facility
type FacilitySuccessRes = SuccessResponseWithData[Facility]
type FacilityListRes = SuccessResponseWithMeta[[]Facility]